Refactor some common utilities to utils, remove print statements and debugging code
