+++ espei/plot.py
@@ -15,7 +15,8 @@ from pycalphad.plot.eqplot import eqplot, _map_coord_to_variable, unpack_conditi
 
 from espei.utils import bib_marker_map
 from espei.core_utils import get_data, get_samples, list_to_tuple, \
-    endmembers_from_interaction, build_sitefractions, ravel_zpf_values, ravel_conditions
+    ravel_zpf_values, ravel_conditions
+from espei.utils import endmembers_from_interaction, build_sitefractions
 
 plot_mapping = {
     'T': 'Temperature (K)',
